requirements.txt
openai==1.51.0
faiss-cpu==1.8.0.post1
numpy==1.26.4
langchain==0.3.1
Janome==0.5.0
rank-bm25==0.2.2
Flask==3.0.3
jsonlines==4.0.0